requirements.txt
streamlit==1.9.0
pandas==1.3.3
numpy==1.21.1
tensorflow==2.6.0
Pillow==8.2.0
pymongo==3.11.4
